moved core dependency into */Gemfile with :path
In order to get the rails generators and the like to work properly in engines we need to require all the dependencies in the engine.rb file.

Since I want to keep that list of engines in a centralized place we still need core and we need to require it from the other engines.

We don't want to require the core gem to be installed so I added it with :path option to the Gemfile.