remove test for duplicate login - we'll prevent that on the client side
SRP happens in two steps:
* handshake
* validation

During the validation we delete the handshake data from the session. So a second validation does not really work. It could build upon the first one but it would not be able to send M2 to the client.

So instead of trying to do sth. usefull when two validation requests are send we require the client to only send one.