add dump_design_docs to CouchRest::Model::Utils:Migrate
This is similar to the migrations but instead of uploading the design
documents to couch it stores them in
  tmp/database/design.json
within the rails directory.

database is the supposed database name without prefixes or suffixes
design is the name of the design doc CouchRest model would have created

The files also contain a couchrest checksum so couchrest can detect they
are up to date.

This commit also cleans up a few redundant things in the extension to
CouchRest::Model:Utils::Migrate that we used to have. There's no need
to loop through the 'normal' models in load_all_models_with_engines
since load_all_models_without_engines already does that. We were also
overwriting all_models_and_proxies with exactly the same code as in
the original.