refactor: move nagios specifs to nagios_test
nagios_test.run takes a function and executes it.
If it returns nothing or 0 and OK nagios message is printed.
If it returns sth. else this will be printed a a warning
If it raises an exception that will result in a CRITICAL report.

This way we can keep the nagios things outside the test cases and just write simple functions that either return 0, a warnign or raise a meaningful exception