return nil as auto_ticket_path for invalid tickets
The auto_ticket_path is referenced from the tickets_controller like this:
     respond_with(@ticket, :location => auto_ticket_path(@ticket))
While this only uses the location parameter when the ticket is valid it will still attampt to calculate it if not. During the create action this will lead to crashes because the ticket_path can not be calculated for a ticket without an id.

This led to https://leap.se/code/issues/5552 and probably
https://leap.se/code/issues/5545.