basic password validation without client side gem
The client_side_validations gem is not maintained anymore and the validations
were not working lately. So instead of trying to fix it I started working on
independent validations for the password as it can't be validated on the
server due to SRP.

So far these validations are very primitive. They require 8 characters length
and a matching confirmation.