prefill ticket form from the model - fixes #5657
email and regarding user fields can be set to defaults based on
created_by user.

If these fields are emptied by the submitting user they will be set to  whereas they are nil if they have not been initialized. In that case we will use meaningful defaults from the user who created the ticket.