move User Control Panel heading out of masthead
the masthead can only handle content of a limited width.

Alternatively we could make the masthead title wider and add padding on the left so it still aligns nicely with the mask. However the wider we make it the worse it looks on small width displays.

Another option would be to make the masthead contain multiple lines. However vertical aligment to the middle of 1 OR 2 lines of text is a real pain.

So I went with a super simple masthead and the Caption goes below.