split up and refactor TicketController#update
close and open actions for plain opening and closing the tickets
respond_with so fields are not cleared on invalid update

the custom actions are not strictly restful. But adding a subresource felt like too much overhead and is conceptually hard to grasp (so we destroy the openess of the ticket to close it?).