remove outdated couchrest hack
We had rewritten use_database so it would not crash the entire app
if couch was not available at the time of initialization.

couchrest now moved on and only sets the database name on use_database.
The database will only be queried once it is really needed. So pretty
much exactly what we want.

Our hack instead caused quite a bit of problems as it would still
initialize the database during start. This way the app would remain
in an invalid state even when couch came back.