Fix db:migrate and similar tasks
We saw errors from duplicate loading of LocalEmail and
LoginFormatValidation. The latter resulted in a crash.

In an attempt to ensure all subclasses of Couchrest::Model::Base are
loaded Couchrest::Model::Utils::Migrate requires all files in app/models.
We have an extension that does the same for the engines.

During this process LoginFormatValidation and LocalEmail were autoloaded
when 'identity' was required. Afterwards they were required again.
It looks like rails' autoload mechanism does not play nicely with require.

So to make sure they are not autoloaded first move the concerns and
helper classes into the lib directory and require them explicitly.