[db] remove DatabaseMethod import from TemporaryUser
It was causing 404s on initializing the database with db:migrate.
DatabaseMethods rewrite of #database does not call prepare_database
and thus does not create the database when needed.

DatabaseMethod is also quite a bit too much for a model that has
two databases that can be memoized easily. There's no way to have
a separate database instance per record.