backport: new customization strategy for the provider message
We display a message for some providers like 'This is a demonstration provider.
Accounts may be deleted at any time.'

We used to do so by just customizing the entire home/content partial. This
had the downside that changes to this partial would not be reflected on the
customized providers.

Instead we now have a separate partial for that purpose.