fix: display error js responses in dev environment
Sometimes the dev environment will send back a plain text response.
This causes the json parser to raise an exception and used to cause
the browser not to display any error message.

Now we dumpt the whole server response - which happes to also include the
backtrace. A lot more useful than doing nothing.

In production this should never happen as 500s get handled by the
ExceptionApplication / our ErrorsController there.