fix: make customization available to sass
Somehow sass did not follow the rails assets path order.
Therefore the default tail.scss would stay in effect even when
there was a different tail.scss in the customization folder.

Adding the customization stylesheet folder to the sass load_paths
works around this for now. Still need to check if it works in
production though.